+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/oras-project/oras-py/tree/main) (0.0.x)
+ - ensure we always strip path separators before pull/push (0.1.12)
  - exposing download_blob to the user since it uses streaming (0.1.11)
    - adding developer examples for pull.
    - start deprecation for _download_blob, _put_upload, _chunked_upload, _upload_manifest